refactor(types): move governance types into its file
